requirements.txt
streamlit
streamlit-option-menu
streamlit-chat
streamlit-webrtc
streamlit-ace
langchain
langchain_community
langchain-google-genai
python-dotenv
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
SpeechRecognition
pyaudio
vosk
sounddevice
numpy
opencv-python-headless
mediapipe
python-dotenv
streamlit-modal
google-auth-oauthlib
google-generativeai
google.ai.generativelanguage
langchain
langchain_community
langchain-google-genai
python-dotenv
langchain_experimental
sentence-transformers
langchainhub
pypdf
rapidocr-onnxruntime
langchain_chroma
pydub
audio-recorder-streamlit